feat(taskboard): Phase D+E — physical delete legacy_store + thread_classifier move + tasks vacuum alias
GAP HIGH closure (audit lines 11-30): legacy_store.py is now gone. The
sidecar JSON write path (~/.frago/ingested_tasks.json) no longer exists in
src/. board.timeline.jsonl is the only persistence layer per spec
20260512 v1.2 freeze.

GAP MID #1 (audit line 32, spec line 67): thread_classifier moves into the
taskboard package — it is a board-attribution concern, not a generic
service.

GAP MID #2 (audit line 33, spec line 11): `frago task vacuum` alias added
forwarding into the shared _run_vacuum implementation from timeline_commands.
The original `frago timeline vacuum` remains for compatibility.

GAP MID #3 (audit line 34): four dead tests removed (they import the
already-deleted thread_service module and would error pytest collection):
- tests/unit/server/services/test_thread_service.py
- tests/unit/server/services/test_reflection_tick.py
- tests/unit/server/services/test_timeline_thread_context.py
- tests/unit/server/services/test_thread_classifier.py

Other:
- git rm src/frago/server/services/taskboard/legacy_store.py
- git mv src/frago/server/services/thread_classifier.py
  → src/frago/server/services/taskboard/thread_classifier.py
- Caller imports updated (task_lifecycle.py + ingestion/scheduler.py +
  test_repeat_response_production_path.py).
- execution_applier.py docstring cleaned of "Phase 1 后续" cruft now that
  executor is integrated.
- decision_applier.py SIM102 fix (combine nested if).
- timeline.py UP035 + SIM117 fixes (Iterator from collections.abc; merged
  with-statements).
- test_pa_validators.py: VALID_PA_ACTIONS expectation realigned with code
  ("dismiss" not "schedule") — pre-existing drift.

Verify:
- ruff check src/frago/{server/services/taskboard,server/services/executor.py,
  server/services/task_lifecycle.py,server/services/primary_agent_service.py,
  server/services/ingestion,server/app.py,server/daemon.py,server/routes/pa.py,
  cli/task_commands.py,cli/timeline_commands.py,server/services/trace.py}
  → All checks passed.
- pytest tests/unit/ tests/server/ → 746 passed, 4 skipped (Windows-only).
