+++ src/frago/server/services/taskboard/execution_applier.py
@@ -1,10 +1,11 @@
 """ExecutionApplier: Executor 启动 / 完成回写 board.
 
-Phase 1 接口:
+接口:
 - start_task(task_id, run_id, pid, csid) — Executor spawn 后调
 - finish_task(task_id, result, error) — sub-agent 完成 / 失败时调
 
-Phase 1 后续 commit: executor.py 集成 (从 TaskStore 改调本类)。
+Spec 20260512 v1.2 freeze: board.timeline.jsonl 是唯一持久化源,
+executor 直接调本类(及 board public methods)无 TaskStore 旁路.
 """
 
 from __future__ import annotations
